moved user query to query folder
