+++ db/queries/users.js
@@ -7,4 +7,18 @@ const getUsers = () => {
     });
 };
 
-module.exports = { getUsers };
+const getUserWithEmail = function(email) {
+  return db
+  .query(`
+  SELECT *
+  FROM users
+  WHERE users.email = $1;
+  `,[email])
+  .then(res => {
+    console.log(res.rows);
+    return res.rows[0] || null;
+  })
+  .catch(err => console.error(err.message));
+};
+
+module.exports = { getUsers, getUserWithEmail };
